requirements.txt
mkdocs
mkdocs-material
pyyaml
jsonschema
python-slugify
spacy
